+++ asyncpraw/models/reddit/rules.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, SupportsIndex
 from urllib.parse import quote
 
 from asyncpraw.const import API_PATH
@@ -136,7 +136,7 @@ class RuleModeration:
         .. code-block:: python
 
             subreddit = await reddit.subreddit("test")
-            rule = await subreddit.rules.get_rule("No Spam")
+            rule = await subreddit.rules.get_rule("No Spam", fetch=False)
             await rule.mod.delete()
 
         """
@@ -284,10 +284,12 @@ class SubredditRules:
             rule.subreddit = self.subreddit
         return rule_list
 
-    async def get_rule(self, short_name: str | (int | slice)) -> asyncpraw.models.Rule:
-        """Return the :class:`.Rule` for the subreddit with short_name ``short_name``.
+    async def get_rule(self, /, short_name: SupportsIndex, *, fetch: bool = True) -> asyncpraw.models.Rule:
+        """Return the :class:`.Rule` for the subreddit with the short name/number/slice ``short_name``.
 
-        :param short_name: The short_name of the rule, or the rule number.
+        :param short_name: The short name of the rule, or the rule number.
+        :param fetch: Determines if Async PRAW will fetch the object (default:
+            ``True``).
 
         This method is to be used to fetch a specific rule, like so:
 
@@ -318,12 +320,31 @@ class SubredditRules:
             subreddit = await reddit.subreddit("test")
             rule = await subreddit.rules.get_rule(1)
 
+        To get the last three rules in a subreddit:
+
+        .. code-block:: python
+
+            subreddit = await reddit.subreddit("test")
+            rules = await subreddit.rules.get_rule(slice(-3, None))
+            for rule in rules:
+                print(rule)
+
+        If you don't need the object fetched right away (e.g., to utilize a class
+        method) you can do:
+
+        .. code-block:: python
+
+            subreddit = await reddit.subreddit("test")
+            rule = await subreddit.rules.get_rule("No Spam", fetch=False)
+            await rule.mod.delete()
+
         """
         if not isinstance(short_name, str):
             rules = await self._rule_list()
             return rules[short_name]
         rule = Rule(self._reddit, subreddit=self.subreddit, short_name=short_name)
-        await rule._fetch()
+        if fetch:
+            await rule._fetch()
         return rule